Fix everything but the rendering stuff.
Update Tiles to use the new read method which has a blockstate parameter.
Fix the lang provider overriding a biomes method that no longer exists
Fix the Gui Helper which was using a double when a float was requred for uvs
Fix TextureHelper which was using a method that no longer exists for getting a TextureAtlasSprite
Fix the VanillaMessageDispatcher using fields rather than the encapsulation methods.
Delete proxy, proxies are no longer required due to the sided setup methods.
Update mods.toml and pack.mcmeta to reflect new versions of things
Setup a testmod and remove all testing code from the actual mod.

Signed-off-by: Luke Gilfoyle <boo122333@gmail.com>